tests/wpt/tests/css/css-view-transitions/navigation/pageswap-push-from-click.html

<!DOCTYPE html>
<title>View transitions: pageswap navigationactivation for push navigations from user click</title>
<link rel="help" href="https://drafts.csswg.org/css-view-transitions-2/">
<link rel="author" href="mailto:khushalsagar@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<style>
@view-transition {
  navigation: auto;
}
</style>
<script>
const expectedUrl = location.href + '?new';

const params = new URLSearchParams(location.search);
// The page the popup navigates to.
const is_new_page = params.has('new');
// The initial page in the popup.
const is_popup_page = params.has('popup');
// The test page itself.
const is_test_page = !is_popup_page && !is_new_page;

const channel = new BroadcastChannel("testchannel");

if (is_test_page) {
  const expectedUrl = location.href + "?new";
  const expectedEvents = ["pageswap", "transition", expectedUrl, "push","from", "pagehide"];

  promise_test(async t => {
    let popup;
    onload = () => {
      document.getElementById('nav_link').remove();
      window.events = [];
      popup = window.open("?popup");

      popup.addEventListener("load", () => {
        popup.requestAnimationFrame(
          () => popup.requestAnimationFrame(() => {
            let nav_link = popup.document.getElementById('nav_link');
            test_driver
                .click(nav_link)
                .catch(() => assert_unreached("click failed"));
          }));
      });
    };

    await new Promise(resolve => {
      channel.addEventListener(
        "message", t.step_func(async (e) => {
          if (e.data === "nav") {
            assert_array_equals(window.events, expectedEvents, 'incorrect event order');
            popup.close();
            resolve();
          }
      }));
    });
  }, `pageswap on navigation from user click`);
} else if (is_popup_page) {
  onpageswap = (e) => {
    window.opener.events.push("pageswap");
    if (e.viewTransition != null)
      window.opener.events.push("transition");
    window.opener.events.push(e.activation.entry.url);
    window.opener.events.push(e.activation.navigationType);
    if (e.activation.from == navigation.currentEntry)
      window.opener.events.push("from");
  };

  onpagehide = () => {
    window.opener.events.push("pagehide");
    channel.postMessage("nav");
  };
}
</script>
<body>
  <a id="nav_link" href='/css/css-view-transitions/navigation/pageswap-push-from-click.html?new'>Click me</a>
 </body>
